requirements.txt
requests
sqlalchemy
psycopg2-binary
apache-airflow==2.10.5